requirements.txt
Flask==2.2.3
MarkupSafe==2.1.2
numpy==1.21.5
Pillow==9.2.0
tensorflow==2.11.0
Jinja2==3.1.2




